@yihuang 2023-03-14
> runMsgs seems heavier than ante handles

that's allocation?