@mmsqe 2023-03-14
> > runMsgs seems heavier than ante handles
> 
> that's allocation?

yes, and cpuprofile shows
<img width="709" alt="Screenshot 2023-03-14 at 4 54 24 PM" src="https://user-images.githubusercontent.com/2867936/224947667-b2c7b45d-ceab-4150-8f27-aa25f36d0944.png">