@yihuang 2023-03-14
> > > runMsgs seems heavier than ante handles
> > 
> > 
> > that's allocation?
> 
> yes, and cpuprofile shows <img alt="Screenshot 2023-03-14 at 4 54 24 PM" width="709" src="https://user-images.githubusercontent.com/2867936/224947667-b2c7b45d-ceab-4150-8f27-aa25f36d0944.png">

ok, they are close enough, the result is not 100% reproducible.